First set of updates for DASH

+++ app/src/main/java/com/breadwallet/tools/security/PassCodeManager.java
@@ -55,11 +55,11 @@ public class PassCodeManager {
         setSpendingLimitIfNotSet(context);
     }
 
-    public int getLimit(Activity activity) {
+    public long getLimit(Activity activity) {
         return SharedPreferencesManager.getLimit(activity);
     }
 
-    public void setLimit(Activity activity, int limit) {
+    public void setLimit(Activity activity, long limit) {
         SharedPreferencesManager.putLimit(activity, limit);
     }
